Initial implementation of tiling and other stuff
Started to have a default implementation for tiling using chunks and
culling in order to reduce the number of draw calls, especially for
objects that aren't visible. Later I'll need to update the material
to allow for the dynamic lighting. Also need to make sure that tiles
get mapped to their correct world space coordinates, as that is the
current roadblock.