Added support for multiple tile spanning objects
Can now place multiple tile spanning objects. All the logic is now
placed in MapReader.js, which is probably not the best place for it.
Also made some major changes to the TimeOfDay system, and a lot more
controls. The sun had been lighting up the undersides and front side
of the PokeCenter when the rest of the world was black, so now there
are functions and controls in place to control the color of the light
throughout the dat